+++ src/prime_rl/trainer/models/layers/lora/__init__.py
@@ -5,9 +5,9 @@ from prime_rl.trainer.models.layers.lora.base import (
     set_lora_num_tokens,
     set_multilora_scaling,
 )
-from prime_rl.trainer.models.layers.lora.fused_moe import MultiLoRAPERFTE
 from prime_rl.trainer.models.layers.lora.multi_linear import MultiLoRALinear
 from prime_rl.trainer.models.layers.lora.multi_moe import MultiLoRAGroupedExperts
+from prime_rl.trainer.models.layers.lora.multi_perfte import MultiLoRAPERFTE
 
 __all__ = [
     "MultiLoRAModule",
